+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 ﻿# @addfox/core
 
+## 0.2.6
+
+### Patch Changes
+
+- chore: update cache and config
+
 ## 0.2.5
 
 ### Patch Changes