+++ packages/plugins/rsbuild-plugin-extension-hmr/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @addfox/rsbuild-plugin-extension-hmr
 
+## 0.2.6
+
+### Patch Changes
+
+- chore: update cache and config
+- Updated dependencies
+  - @addfox/launcher@0.2.6
+  - @addfox/core@0.2.6
+
 ## 0.2.5
 
 ### Patch Changes